+++ R/rhat.R
@@ -1,6 +1,6 @@
 #' R-hat
 #'
-#' By default calculates the uncorrected, untransformed, univariate
+#' By default calculates the uncorrected, univariate, untransformed,
 #' split R-hat (potential scale reduction factor) values.
 #'
 #' @param x An MCMC object.
@@ -9,7 +9,8 @@
 #' "term", "parameter" or "all".
 #' @param as_df A flag indicating whether to return the values as a
 #' data frame versus a named list.
-#' @param split A flag specifying whether to split the chains in two.
+#' @param split A flag specifying whether to split the chains.
+#' @param normalized A flag specifying whether to rank normalize the values.
 #' @param bound flag specifying whether to bind mcmcrs objects by their chains before calculating rhat.
 #' @return The rhat value(s).
 #' @references
@@ -28,32 +29,40 @@ rhat <- function(x, ...) {
 
 #' @describeIn rhat R-hat for an mcarray object
 #' @export
-rhat.mcarray <- function(x, by = "all", as_df = FALSE, split = TRUE, ...) {
+rhat.mcarray <- function(x, by = "all", as_df = FALSE,
+                         split = TRUE, normalized = FALSE, ...) {
   check_unused(...)
-  rhat(as.mcmcarray(x), by = by, as_df = as_df, split = TRUE)
+  rhat(as.mcmcarray(x), by = by, as_df = as_df,
+       split = split, normalized = normalized)
 }
 
 #' @describeIn rhat R-hat for an mcmc object
 #' @export
-rhat.mcmc <- function(x, by = "all", as_df = FALSE, split = TRUE, ...) {
+rhat.mcmc <- function(x, by = "all", as_df = FALSE,
+                      split = TRUE, normalized = FALSE, ...) {
   check_unused(...)
-  rhat(as.mcmcr(x), by = by, as_df = as_df, split = split)
+  rhat(as.mcmcr(x), by = by, as_df = as_df,
+       split = split, normalized = normalized)
 }
 
 #' @describeIn rhat R-hat for an mcmc.list object
 #' @export
-rhat.mcmc.list <- function(x, by = "all", as_df = FALSE, split = TRUE, ...) {
+rhat.mcmc.list <- function(x, by = "all", as_df = FALSE, split = TRUE,
+                           normalized = FALSE, ...) {
   check_unused(...)
-  rhat(as.mcmcr(x), by = by, as_df = as_df, split = split)
+  rhat(as.mcmcr(x), by = by, as_df = as_df,
+       split = split, normalized = normalized)
 }
 
 #' @describeIn rhat R-hat for an mcmcarray object
 #' @export
-rhat.mcmcarray <- function(x, by = "all", as_df = FALSE, split = TRUE, ...) {
+rhat.mcmcarray <- function(x, by = "all", as_df = FALSE,
+                           split = TRUE, normalized = FALSE, ...) {
   check_unused(...)
   check_vector(by, c("all", "parameter", "term"), length = 1)
   check_flag(as_df)
   check_flag(split)
+  check_flag(normalized)
 
   if(niters(x) < 4) {
     if(!as_df) {
@@ -66,7 +75,7 @@ rhat.mcmcarray <- function(x, by = "all", as_df = FALSE, split = TRUE, ...) {
   }
 
   if(split) x <- split_chains(x)
-  x <- apply(x, 3:ndims(x), FUN = .rhat)
+  x <- apply(x, 3:ndims(x), FUN = .rhat, normalized = normalized)
 
   if(!as_df) {
     if(by == "term") return(x)
@@ -81,10 +90,12 @@ rhat.mcmcarray <- function(x, by = "all", as_df = FALSE, split = TRUE, ...) {
 
 #' @describeIn rhat R-hat for an mcmcr object
 #' @export
-rhat.mcmcr <- function(x, by = "all", as_df = FALSE, split = TRUE, ...) {
+rhat.mcmcr <- function(x, by = "all", as_df = FALSE,
+                       split = TRUE, normalized = FALSE, ...) {
   check_unused(...)
   parameters <- parameters(x)
-  x <- lapply(x, rhat, by = by, as_df = as_df, split = split)
+  x <- lapply(x, rhat, by = by, as_df = as_df,
+              split = split, normalized = normalized)
   if(!as_df) {
     if (by != "all") return(x)
     return(max(unlist(x)))
@@ -98,14 +109,17 @@ rhat.mcmcr <- function(x, by = "all", as_df = FALSE, split = TRUE, ...) {
 
 #' @describeIn rhat R-hat for an mcmcrs object
 #' @export
-rhat.mcmcrs <- function(x, by = "all", as_df = FALSE, split = !bound, bound = FALSE, ...) {
+rhat.mcmcrs <- function(x, by = "all", as_df = FALSE,
+                        split = !bound, bound = FALSE, normalized = FALSE, ...) {
   check_flag(bound)
   check_unused(...)
 
   if(bound) {
     x <- lapply(x, collapse_chains)
     x <- Reduce(bind_chains, x)
-    return(rhat(x, by = by, as_df = as_df, split = split))
+    return(rhat(x, by = by, as_df = as_df,
+                split = split, normalized = normalized))
   }
-  lapply(x, rhat, by = by, as_df = as_df, split = split)
+  lapply(x, rhat, by = by, as_df = as_df,
+         split = split, normalized = normalized)
 }