+++ NEWS.md
@@ -1,3 +1,4 @@
+- add normalized = FALSE argument to rhat and converged
 - add split = TRUE argument to rhat and converged
 - define subset and parameters for mcmcrs object
 - add bound = FALSE argument to rhat.mcmcrs and converged.mcmcrs